refactor: split impl of x86_64 specific code
